refactor: apply components folder
