+++ apps/web/src/services/dashboards/components/dashboard-create/DashboardCreateTemplateBoard.vue
