+++ apps/web/src/services/dashboards/components/dashboard-detail/DashboardControlButtons.vue
