+++ apps/web/src/services/dashboards/components/dashboard-detail/DashboardLabels.vue
