+++ apps/web/src/services/dashboards/components/dashboard-detail/DashboardReorderSidebar.vue
