+++ apps/web/src/services/dashboards/components/dashboard-detail/DashboardVariableDropdown.vue
