+++ apps/web/src/services/dashboards/components/legacy/DashboardCostWidgetValueOptionDropdown.vue
