+++ apps/web/src/services/dashboards/components/legacy/DashboardManageVariableOptionsField.vue
