+++ apps/web/src/services/dashboards/components/legacy/DashboardManageVariableTable.vue
