+++ apps/web/src/services/dashboards/components/legacy/DashboardVariablesMoreButton.vue
