+++ apps/web/src/services/dashboards/components/legacy/DashboardWidgetMoreOptions.vue
