+++ apps/web/src/services/dashboards/pages/DashboardDetailPage.vue
@@ -23,13 +23,13 @@ import { FAVORITE_TYPE } from '@/common/modules/favorites/favorite-button/type';
 import { useGnbStore } from '@/common/modules/navigations/stores/gnb-store';
 import { useWidgetGenerateStore } from '@/common/modules/widgets/_store/widget-generate-store';
 
-import DashboardDetailHeader from '@/services/dashboards/components/DashboardDetailHeader.vue';
-import DashboardRefreshDropdown from '@/services/dashboards/components/DashboardRefreshDropdown.vue';
-import DashboardToolsetDateDropdown from '@/services/dashboards/components/DashboardToolsetDateDropdown.vue';
-import DashboardVariables from '@/services/dashboards/components/DashboardVariables.vue';
-import DashboardVariablesV2 from '@/services/dashboards/components/DashboardVariablesV2.vue';
-import DashboardWidgetContainer from '@/services/dashboards/components/DashboardWidgetContainer.vue';
-import DashboardWidgetContainerV2 from '@/services/dashboards/components/DashboardWidgetContainerV2.vue';
+import DashboardDetailHeader from '@/services/dashboards/components/dashboard-detail/DashboardDetailHeader.vue';
+import DashboardRefreshDropdown from '@/services/dashboards/components/dashboard-detail/DashboardRefreshDropdown.vue';
+import DashboardToolsetDateDropdown from '@/services/dashboards/components/dashboard-detail/DashboardToolsetDateDropdown.vue';
+import DashboardVariablesV2 from '@/services/dashboards/components/dashboard-detail/DashboardVariablesV2.vue';
+import DashboardWidgetContainerV2 from '@/services/dashboards/components/dashboard-detail/DashboardWidgetContainerV2.vue';
+import DashboardVariables from '@/services/dashboards/components/legacy/DashboardVariables.vue';
+import DashboardWidgetContainer from '@/services/dashboards/components/legacy/DashboardWidgetContainer.vue';
 import { DASHBOARD_SCOPE } from '@/services/dashboards/constants/dashboard-constant';
 import { DASHBOARD_TEMPLATES } from '@/services/dashboards/dashboard-template/template-list';
 import { DASHBOARDS_ROUTE } from '@/services/dashboards/routes/route-constant';
